backend/requirements.txt
django-cors-headers
djangorestframework
django-summernote
django-rest-knox
django-seed
django-filter
